Revert changing dimensions from int[] to long[]. We decided that a single dimension cannot be larger than Integer.MAX_VALUE for now, mostly because most algorithms that are linked and used transfer at least a single pixel row into a Java Array by which this limitation is dictated.  If, at some point, this decision constrains us, we can change back.

+++ mpicbg/imglib/container/ContainerFactory.java
@@ -44,7 +44,7 @@ public abstract class ContainerFactory implements Factory
 	 * 
 	 * @return {@link Container} - the instantiated Container
 	 */
-	public abstract <T extends Type<T>> Container<T> createContainer( final long[] dim, final T type );
+	public abstract <T extends Type<T>> Container<T> createContainer( final int[] dim, final T type );
 	
 	public void setOptimizedContainerUse ( final boolean useOptimizedContainers ) { this.useOptimizedContainers = useOptimizedContainers; }
 	public boolean useOptimizedContainers() { return useOptimizedContainers; }
