+++ mpicbg/imglib/container/dynamic/BitDynamicContainer.java
@@ -40,7 +40,7 @@ public class BitDynamicContainer <T extends Type<T>> extends DynamicContainer<T,
 	final static int bitsPerEntity = Integer.SIZE;
 	final ArrayList<Integer> data;
 	
-	public BitDynamicContainer( final DynamicContainerFactory factory, final long[] dim, final int entitiesPerPixel )
+	public BitDynamicContainer( final DynamicContainerFactory factory, final int[] dim, final int entitiesPerPixel )
 	{
 		super( factory, dim, entitiesPerPixel );
