+++ mpicbg/imglib/container/dynamic/DoubleDynamicContainer.java
@@ -39,7 +39,7 @@ public class DoubleDynamicContainer <T extends Type<T>> extends DynamicContainer
 {
 	final ArrayList<Double> data;
 	
-	public DoubleDynamicContainer( final DynamicContainerFactory factory, final long[] dim, final int entitiesPerPixel )
+	public DoubleDynamicContainer( final DynamicContainerFactory factory, final int[] dim, final int entitiesPerPixel )
 	{
 		super( factory, dim, entitiesPerPixel );
