+++ mpicbg/imglib/container/imageplus/ShortImagePlus.java
@@ -40,7 +40,7 @@ public class ShortImagePlus<T extends Type<T>> extends ImagePlusContainer<T, Sho
 {
 	final ImagePlus image;	
 	
-	public ShortImagePlus( final ImagePlusContainerFactory factory, final long[] dim, final int entitiesPerPixel ) 
+	public ShortImagePlus( final ImagePlusContainerFactory factory, final int[] dim, final int entitiesPerPixel ) 
 	{
 		super( factory, dim, entitiesPerPixel );
